requirements.txt
requests
pycryptodome
curl_cffi
aiohttp
certifi
browser_cookie3
websockets
pyexecjs
pytz
openai
python-dotenv
websocket-client
tls-client
free-proxy
fastapi[all]==0.100.0
slack-sdk==3.21.3
poe-api==0.4.17
vercel-llm-api==0.3.0
beautifulsoup4==4.12.2